# web/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Iris API",
    "description": "Self-hosted media aggregator + streamer — internal household API.",
    "license": {
      "name": "MIT OR Apache-2.0",
      "identifier": "MIT OR Apache-2.0"
    },
    "version": "0.8.0"
  },
  "paths": {
...
    "/api/me/playback-preferences": {
      "get": {
        "tags": [
          "preferences"
        ],
        "operationId": "get_playback_preferences",
        "responses": {
          "200": {
            "description": "The caller's preferred audio + subtitle languages",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlaybackPrefsResponse"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "preferences"
        ],
        "operationId": "set_playback_preferences",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePlaybackPrefs"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Preferences saved"
          }
        }
      }
    },
    "/api/me/preferences": {
      "get": {
        "tags": [
          "preferences"
        ],
        "operationId": "get_preferences",
        "responses": {
          "200": {
            "description": "The caller's recommendation preferences",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PreferencesResponse"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "preferences"
        ],
        "operationId": "update_preferences",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePreferencesRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Saved preferences (normalised)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PreferencesResponse"
                }
              }
            }
          },
          "400": {
            "description": "Unknown language in the request"
          }
        }
      }
    },
...
    "/api/metadata/tmdb/resolve": {
      "get": {
        "tags": [
          "metadata"
        ],
        "summary": "Resolve a release name to its single best TMDB match, scored by\n**kind + year** (not raw popularity) and served from the persistent\n30-day resolve cache shared with the ingest/backfill pipeline.",
        "description": "This is the poster-resolution path for search-result cards. It\nreplaces the old client-side `extractSceneTitle` + `/tmdb/search` +\n\"take popularity #1\" logic that was duplicated (and subtly fragile —\nshort titles like \"Pride\" collided with more-popular namesakes) in\nboth the web and TV clients. `/tmdb/search` stays a raw popularity\nproxy for the live typeahead dropdown, which legitimately wants the\nunfiltered suggestion list.",
        "operationId": "resolve_tmdb",
        "parameters": [
          {
            "name": "title",
            "in": "query",
            "description": "Raw SCENE release title, untouched\n(e.g. `Pride.2014.1080p.BluRay.x264-AMIABLE`). The backend parses\ntitle / year / kind out of it — clients send the release name\nverbatim so the cleaning + scoring logic lives in exactly one\nplace instead of being reimplemented per client.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "kind",
            "in": "query",
            "description": "Optional stronger kind hint from the search result's own\nclassification (`movie` | `tv`); overrides the kind parsed from\nthe release name.",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Best TMDB match for the release name (null when none / unconfigured)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TmdbSuggestion"
                }
              }
            }
          }
        }
      }
    },
    "/api/metadata/tmdb/search": {
      "get": {
        "tags": [
          "metadata"
        ],
        "summary": "Typeahead suggestions for the search page. Proxies TMDB's\n`/search/multi` so the API key stays server-side. Empty array on no\nresults, missing TMDB config, or upstream failure (typeahead is best-\neffort: never block typing on a flaky network).",
        "operationId": "search_tmdb",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Typeahead suggestions (empty when TMDB unconfigured)",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TmdbSuggestion"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/metadata/tmdb/{id}": {
      "get": {
        "tags": [
          "metadata"
        ],
        "operationId": "get_tmdb_metadata",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64",
              "minimum": 0
            }
          },
          {
            "name": "kind",
            "in": "query",
            "description": "`\"movie\"` | `\"tv\"`. Without a kind, the lookup tries movie\nfirst then TV — but TMDB uses separate id namespaces, so a\nnumerical id can collide between a movie and an unrelated TV\nshow. Pass the kind whenever the caller knows it\n(collection.kind, search-result kind, etc.) to disambiguate.",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "TMDB metadata for the movie / show",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MediaMetadata"
                }
              }
            }
          },
          "400": {
            "description": "TMDB enrichment is not configured"
          },
          "404": {
            "description": "No TMDB record for this id / kind"
          }
        }
      }
    },
...
    "/api/torrents/{infohash}/files/{idx}/play/{asset}": {
      "get": {
        "tags": [
          "torrents"
        ],
        "summary": "Serve any file from the per-source HLS-CMAF cache directory.",
        "description": "On the master playlist request specifically (`asset == \"master.m3u8\"`)\nwe ensure the cache is built — probe the source, compute the rendition\nplan, and block on the remuxer until the master + first fragment of\nevery variant exist on disk. Subsequent asset requests (variant\nplaylists, init segments, `.m4s`) are pure static-file serving with\nbyte-range support; the player only asks for them after parsing the\nmaster, by which point everything has been observed by the watcher.",
        "operationId": "play_asset",
        "parameters": [
          {
            "name": "infohash",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "idx",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 0
            }
          },
          {
            "name": "asset",
            "in": "path",
            "description": "`master.m3u8`, a variant playlist, an init segment, or an `.m4s` fragment",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "HLS-CMAF asset; `master.m3u8` blocks until ffmpeg has built the head",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "206": {
            "description": "Partial content for an HTTP `Range` request on a segment"
          },
          "404": {
            "description": "Unknown infohash / file / asset"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "MediaMetadata": {
        "type": "object",
        "required": [
          "kind",
          "tmdb_id",
          "title",
          "genres",
          "genre_ids"
        ],
        "properties": {
          "backdrop_path": {
            "type": [
              "string",
              "null"
            ]
          },
          "genre_ids": {
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int32",
              "minimum": 0
            },
            "description": "TMDB genre ids. Always present from list endpoints (discover /\ntrending); derived from the full genre objects on detail lookups."
          },
          "genres": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "kind": {
            "$ref": "#/components/schemas/TmdbKind"
          },
          "number_of_seasons": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32",
            "description": "TV shows only — total seasons published. Used at follow time to\nsnapshot how many seasons we expect, so the Series page can pre-\nrender season tabs without waiting on a fresh TMDB lookup.",
            "minimum": 0
          },
          "original_language": {
            "type": [
              "string",
              "null"
            ],
            "description": "ISO 639-1 original language (\"fr\" / \"en\" / …). Drives per-user\nlanguage filtering of the catalogue."
          },
          "overview": {
            "type": [
              "string",
              "null"
            ]
          },
          "popularity": {
            "type": [
              "number",
              "null"
            ],
            "format": "double",
            "description": "TMDB popularity score (relative, unbounded). Ranks catalogue\ncandidates in the recommendation pipeline."
          },
          "poster_path": {
            "type": [
              "string",
              "null"
            ],
            "description": "Path component, e.g. `/abc.jpg`. Combine with size + base URL to form\nthe actual image URL: `https://image.tmdb.org/t/p/<size><poster_path>`."
          },
          "release_date": {
            "type": [
              "string",
              "null"
            ],
            "description": "`YYYY-MM-DD` release / first-air date, kept raw alongside the\nparsed `year`."
          },
          "runtime_minutes": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32",
            "description": "Movie runtime in minutes (TMDB `runtime`). For TV shows we expose\nthe *typical* episode runtime here when one is published; both\ncan drift wildly from the actual file's duration so callers do\nthe verification themselves.",
            "minimum": 0
          },
          "title": {
            "type": "string"
          },
          "tmdb_id": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          },
          "vote_count": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32",
            "minimum": 0
          },
          "vote_score": {
            "type": [
              "number",
              "null"
            ],
            "format": "double",
            "description": "0..1 (TMDB's `vote_average` is /10, normalized here for the UI)."
          },
          "year": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32",
            "minimum": 0
          }
        }
      },
...
      "PlaybackPrefsResponse": {
        "type": "object",
        "properties": {
          "audio_language": {
            "type": [
              "string",
              "null"
            ],
            "description": "Preferred audio language (ISO 639-1 / BCP-47), or null = no preference."
          },
          "subtitle_language": {
            "type": [
              "string",
              "null"
            ],
            "description": "Preferred subtitle language, `\"off\"` for disabled, or null = no\npreference."
          }
        }
      },
...
      "PreferencesResponse": {
        "type": "object",
        "required": [
          "languages",
          "genres",
          "include_anime",
          "onboarding_completed"
        ],
        "properties": {
          "genres": {
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int64"
            }
          },
          "include_anime": {
            "type": "boolean"
          },
          "languages": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "onboarding_completed": {
            "type": "boolean"
          }
        }
      },
...
      "TmdbKind": {
        "type": "string",
        "enum": [
          "movie",
          "tv"
        ]
      },
      "TmdbSuggestion": {
        "type": "object",
        "required": [
          "kind",
          "tmdb_id",
          "title"
        ],
        "properties": {
          "kind": {
            "$ref": "#/components/schemas/TmdbKind"
          },
          "overview": {
            "type": [
              "string",
              "null"
            ]
          },
          "poster_path": {
            "type": [
              "string",
              "null"
            ]
          },
          "title": {
            "type": "string"
          },
          "tmdb_id": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          },
          "year": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32",
            "minimum": 0
          }
        }
      },
...
      "UpdatePlaybackPrefs": {
        "type": "object",
        "properties": {
          "audio_language": {
            "type": [
              "string",
              "null"
            ]
          },
          "subtitle_language": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
      "UpdatePreferencesRequest": {
        "type": "object",
        "properties": {
          "genres": {
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int64"
            }
          },
          "include_anime": {
            "type": "boolean"
          },
          "languages": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "onboarding_completed": {
            "type": "boolean"
          }
        }
      },
...
    }
  },
...
}